# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Jersey Webshop API
  description: |
    API za webshop fan sportskih dresova (nogomet, F1, košarka).

    Proizvod je jedna tablica: veličine (`sizes`) i slike (`images`) su nizovi
    na samom dresu, uz **jedan** 3D model (`model_3d_url`). Slike se hostaju
    vanjski i šalju kao URL-ovi; prva u nizu je primarna. Zalihe nema — sve
    navedene veličine su uvijek dostupne.

    Ime i broj za tisak upisuje kupac pri narudžbi. Liste igrača backend ne
    poznaje — frontend ih vuče s vanjskog API-ja.

    Admin endpointi traže Sanctum bearer token iz `POST /admin/login`.
  version: 2.0.0
servers:
  - url: https://api.tvojdomena.com/api
    description: Produkcija
  - url: http://localhost:8000/api
    description: Lokalni razvoj (php artisan serve)
...
paths:
...
  /orders/{id}/status:
    get:
      tags: [Orders (public)]
      summary: Status plaćanja samo po referenci narudžbe
      description: >
        Namijenjeno stranici na koju Stripe vrati kupca nakon Checkouta, dok
        čeka webhook potvrdu. Referenca je kratka i pogodiva, pa odgovor
        namjerno ne sadrži ništa osobno — ni email, ni adresu, ni iznos, ni
        tracking. Puni status je iza `/orders/lookup`, koji traži i email.
        Ograničeno na 30 zahtjeva u minuti.
      parameters:
        - name: id
          in: path
          required: true
          description: order_reference, npr. ORD-8F3K2
          schema:
            type: string
          example: ORD-8F3K2
      responses:
        '200':
          description: Status plaćanja
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderPaymentStatus'
        '404':
          description: Narudžba nije pronađena
        '429':
          description: Previše zahtjeva

  # ============================================
  # WEBHOOKS
  # ============================================
...
  /admin/orders/{id}/status:
    put:
      tags: [Admin - Orders]
      summary: Promjena statusa narudžbe (triggira customer notifikaciju)
      security: [{ sanctum: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [ordered, sent_to_supplier, arrived_hr, shipped, delivered]
                tracking_number_internal:
                  type: string
                  description: Popunjava se kad status prelazi u "shipped"
      responses:
        '200':
          description: Status ažuriran. Promjena statusa šalje email notifikaciju kupcu.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderAdmin'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Narudžba nije pronađena
        '422':
          $ref: '#/components/responses/ValidationError'

  # ============================================
  # ADMIN — DASHBOARD
  # ============================================
...
components:
  responses:
    Unauthorized:
      description: Nedostaje ili je nevažeći token
    Forbidden:
      description: Token pripada korisniku koji nije administrator
    ValidationError:
      description: Validacijska greška
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
  securitySchemes:
    sanctum:
      type: http
      scheme: bearer
  schemas:
...
    PaymentStatus:
      type: string
      description: Stanje naplate — zasebna os od fulfillment `status`
      enum: [unpaid, paid, failed, refunded]
    OrderPaymentStatus:
      type: object
      description: >
        Minimalan javni odgovor za lookup samom referencom — bez osobnih
        podataka i bez trackinga.
      properties:
        order_reference:
          type: string
          example: ORD-8F3K2
        payment_status:
          $ref: '#/components/schemas/PaymentStatus'
        status_label:
          type: string
          example: Narudžba zaprimljena
    OrderAdmin:
      type: object
      properties:
        id:
          type: integer
        order_reference:
          type: string
        customer:
          type: object
          properties:
            name:
              type: string
            email:
              type: string
        status:
          type: string
        payment_status:
          $ref: '#/components/schemas/PaymentStatus'
        total_price:
          type: number
        items:
          type: array
          items:
            type: object
            properties:
              product_id:
                type: integer
                description: >
                  Jedino pouzdano razlikovanje dresa — isti naziv (npr.
                  "Dinamo Home 25/26") postoji i u dječjoj i u odrasloj verziji,
                  pod različitim id-em.
              product_name:
                type: string
              size:
                type: string
                description: Veličina koju je kupac naručio, snapshot u trenutku kupnje
              quantity:
                type: integer
              player_name:
                type: string
                nullable: true
                description: Ime za tisak, onako kako ga je kupac upisao; null kad dres ide bez tiska
              player_number:
                type: string
                nullable: true
                description: Broj za tisak; može postojati i bez imena
        created_at:
          type: string
          format: date-time
    ValidationError:
      type: object
      properties:
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
...